refactor: Notion を DataLayer, Domain に分割
