+++ NotionTimer/NotionTimerPackage/Sources/DataLayer/Entity/NotionAuthStatus.swift
@@ -1,0 +1,13 @@
+//
+//  NotionAuthStatus.swift
+//  NotionTimerPackage
+//
+//  Created by Taichi on 2025/04/02.
+//
+
+public enum NotionAuthStatus {
+    case loading
+    case invalidToken
+    case invalidDatabase
+    case complete
+}
