+++ NotionTimer/NotionTimerPackage/Sources/Presentation/Root/DatabaseCreationView.swift
@@ -6,7 +6,8 @@
 //
 
 import SwiftUI
-import Notion
+import DataLayer
+import Domain
 
 struct DatabaseCreationView: View {
     @Environment(NotionService.self) private var notionService
