+++ NotionTimer/NotionTimerPackage/Sources/Presentation/Root/DatabaseSelectionView.swift
@@ -6,7 +6,8 @@
 //
 
 import SwiftUI
-import Notion
+import DataLayer
+import Domain
 
 struct DatabaseSelectionView: View {
     @Environment(NotionService.self) private var notionService: NotionService
