+++ NotionTimer/NotionTimerPackage/Tests/NotionTests/NotionAPIClientTests.swift
@@ -1,5 +1,6 @@
 import Testing
-@testable import Notion
+@testable import DataLayer
+@testable import Domain
 
 import Foundation
